update_image/requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
djangorestframework==3.12.4
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
sorl-thumbnail==12.7.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.2
toml==0.10.2
